module 3.2

+++ Module 3.1/daystoyears.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 int main()
 {
-    int a,b,c;
+    float a,b,c;
     printf("Enter a number to convert the following\n1:Days to years\n2:years to days\nEnter here: ");
-    scanf("%d",&a);
+    scanf("%f",&a);
     if(a==1)
     {
         printf("Enter the number of days: ");
-        scanf("%d",&b);
-        printf("= %d years",b/365);
+        scanf("%f", &b);
+        printf("= %.2f years",b/365);
     }
     else if(a==2)
     {
         printf("Enter the number of years: ");
-        scanf("%d",&c);
-        printf("= %d days",c*365);
+        scanf("%f",&c);
+        printf("= %.2f days",c*365);
     }
     else
     {
